refactor: use single builder model for getting report
