+++ Onspring.API.SDK.Tests/Tests/Unit/Fluent/Reports/GetReportsRequestBuilderTests.cs
@@ -1,0 +1,22 @@
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using NSubstitute;
+using Onspring.API.SDK.Models.Fluent;
+
+namespace Onspring.API.SDK.Tests.Tests.Unit.Fluent
+{
+    [TestClass, ExcludeFromCodeCoverage]
+    public class GetReportsRequestBuilderTests
+    {
+        private static readonly int _appId = 1;
+        private static IOnspringClient _client;
+        private static GetReportsRequestBuilder _builder;
+
+        [ClassInitialize]
+        public static void ClassInit(TestContext testContext)
+        {
+            _client = Substitute.For<IOnspringClient>();
+            _builder = new GetReportsRequestBuilder(_client);
+        }
+    }
+}
