+++ Onspring.API.SDK/Interfaces/Fluent/Reports/IGetReportsByAppRequestBuilder.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using Onspring.API.SDK.Enums;
 using Onspring.API.SDK.Models;
 using Onspring.API.SDK.Models.Fluent;
 
@@ -14,6 +13,6 @@ namespace Onspring.API.SDK.Interfaces.Fluent
         IGetReportsByAppRequestBuilder ForPageNumber(int pageNumber);
         IGetReportsByAppRequestBuilder WithPageSize(int pageNumber);
         Task<ApiResponse<GetReportsForAppResponse>> SendAsync();
-        Task<ApiResponse<GetReportsForAppResponse>> SendAsync(Action<GetReportsByAppRequestBuilderOptions> options);
+        Task<ApiResponse<GetReportsForAppResponse>> SendAsync(Action<GetReportsRequestBuilderOptions> options);
     }
 }
