+++ Onspring.API.SDK/Interfaces/Fluent/Reports/IGetReportsRequestBuilder.cs
@@ -1,0 +1,7 @@
+namespace Onspring.API.SDK.Interfaces.Fluent
+{
+    public interface IGetReportsRequestBuilder
+    {
+        IGetReportsByAppRequestBuilder FromApp(int appId);
+    }
+}
