+++ Onspring.API.SDK/Models/Fluent/Reports/GetReportBuilderOptions.cs
@@ -2,7 +2,7 @@ using Onspring.API.SDK.Enums;
 
 namespace Onspring.API.SDK.Models.Fluent
 {
-    public class GetReportDataRequestBuilderOptions
+    public class GetReportBuilderOptions
     {
         public DataFormat Format { get; set; } = DataFormat.Raw;
         public ReportDataType DataType { get; set; } = ReportDataType.ReportData;
