+++ Onspring.API.SDK/Models/Fluent/Reports/GetReportsRequestBuilderOptions.cs
@@ -2,7 +2,7 @@ using Onspring.API.SDK.Enums;
 
 namespace Onspring.API.SDK.Models.Fluent
 {
-    public class GetReportsByAppRequestBuilderOptions
+    public class GetReportsRequestBuilderOptions
     {
         public int PageNumber { get; set; } = 1;
         public int PageSize { get; set; } = 50;
